feat: Complete DTO alignment for Intelligent Services re-enablement - Add missing fields to FlightTrackingRequest (hexIdent, groundSpeed, lastSeen, track, verticalRate, alert, emergency, spi, isOnGround) - Add missing fields to ShipTrackingRequest (mmsi, heading, navStatus) - Fix AreaSubscriptionRequest field names and add compatibility methods - Add ship-specific methods to WebSocketSubscriptionService - Add lastSeen field to FlightTracking model - Recreate complete IntelligentStorageService with intelligent storage logic - Fix type conversions (Double to Float, String to Integer) - All 28 compilation errors resolved - Build now compiles successfully - Intelligent Services ready for full deployment

+++ src/main/java/com/phamnam/tracking_vessel_flight/controller/AircraftWebSocketController.java
@@ -29,22 +29,21 @@ public class AircraftWebSocketController {
                         SimpMessageHeaderAccessor headerAccessor) {
                 String sessionId = headerAccessor.getSessionId();
                 log.info("Client {} đăng ký khu vực: {},{} đến {},{}",
-                        sessionId, request.getMinLat(), request.getMinLon(),
-                        request.getMaxLat(), request.getMaxLon());
+                                sessionId, request.getMinLat(), request.getMinLon(),
+                                request.getMaxLat(), request.getMaxLon());
 
                 // Thêm log chi tiết
                 log.debug("STOMP headers: {}", headerAccessor.getMessageHeaders());
                 log.debug("Processing area subscription with sessionId: {}", sessionId);
 
                 try {
-                        // Gọi service với đầy đủ log
+                        // Gọi service với đầy đủ log - convert Double to float
                         subscriptionService.subscribeToArea(
-                                sessionId,
-                                request.getMinLat(),
-                                request.getMaxLat(),
-                                request.getMinLon(),
-                                request.getMaxLon()
-                        );
+                                        sessionId,
+                                        request.getMinLat().floatValue(),
+                                        request.getMaxLat().floatValue(),
+                                        request.getMinLon().floatValue(),
+                                        request.getMaxLon().floatValue());
                         log.info("Area subscription processed successfully for {}", sessionId);
                 } catch (Exception e) {
                         log.error("Error processing area subscription: {}", e.getMessage(), e);
@@ -64,8 +63,8 @@ public class AircraftWebSocketController {
                                 request.getMaxLat(), request.getMaxLon());
 
                 subscriptionService.unsubscribeFromArea(sessionId,
-                                request.getMinLat(), request.getMaxLat(),
-                                request.getMinLon(), request.getMaxLon());
+                                request.getMinLat().doubleValue(), request.getMaxLat().doubleValue(),
+                                request.getMinLon().doubleValue(), request.getMaxLon().doubleValue());
         }
 
         /**
